+++ app/src/main/java/ru/netology/nmedia/viewmodel/PostViewModel.kt
@@ -1,0 +1,13 @@
+package ru.netology.nmedia.viewmodel
+
+import androidx.lifecycle.ViewModel
+import ru.netology.nmedia.repository.PostRepository
+import ru.netology.nmedia.repository.PostRepositoryInMemoryImplementation
+
+class PostViewModel: ViewModel() {
+
+    private val repository: PostRepository = PostRepositoryInMemoryImplementation()
+
+    val data = repository.get()
+    fun like() = repository.like()
+}
